Dedupe some assistant related code logic
