+++ frontend/ui/src/features/ai-assistant/components/ai-chat-overlay.tsx
@@ -2,13 +2,13 @@
 
 import { X, Plus, History, Square } from "lucide-react";
 import { useQuery } from "@tanstack/react-query";
-import { useCallback, useRef, useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
 import { MessageList } from "./message-list";
 import { MessageInput } from "./message-input";
 import { SessionHistory } from "./session-history";
 import { useAiChat } from "../hooks/use-ai-chat";
+import { usePanelResize } from "../hooks/use-panel-resize";
 import { getProject } from "@/lib/api";
 
 interface AiChatOverlayProps {
@@ -21,38 +21,8 @@ interface AiChatOverlayProps {
  * AI chat panel for use inside trace detail viewer.
  * traceId is passed through to the agent so it knows which trace the user is viewing.
  */
-const MIN_WIDTH = 280;
-const MAX_WIDTH = 900;
-const DEFAULT_WIDTH = 400;
-
 export function AiChatOverlay({ projectId, traceId, onClose }: AiChatOverlayProps) {
-  const [width, setWidth] = useState(DEFAULT_WIDTH);
-  const dragging = useRef(false);
-  const startX = useRef(0);
-  const startWidth = useRef(0);
-
-  const onMouseDown = useCallback(
-    (e: React.MouseEvent) => {
-      dragging.current = true;
-      startX.current = e.clientX;
-      startWidth.current = width;
-      e.preventDefault();
-
-      const onMouseMove = (e: MouseEvent) => {
-        if (!dragging.current) return;
-        const delta = startX.current - e.clientX;
-        setWidth(Math.min(MAX_WIDTH, Math.max(MIN_WIDTH, startWidth.current + delta)));
-      };
-      const onMouseUp = () => {
-        dragging.current = false;
-        window.removeEventListener("mousemove", onMouseMove);
-        window.removeEventListener("mouseup", onMouseUp);
-      };
-      window.addEventListener("mousemove", onMouseMove);
-      window.addEventListener("mouseup", onMouseUp);
-    },
-    [width],
-  );
+  const { width, onMouseDown } = usePanelResize();
 
   const { data: project } = useQuery({
     queryKey: ["project", projectId],
