+++ frontend/ui/src/features/ai-assistant/constants.ts
@@ -1,0 +1,3 @@
+export const PANEL_MIN_WIDTH = 280;
+export const PANEL_MAX_WIDTH = 900;
+export const PANEL_DEFAULT_WIDTH = 400;
