+++ frontend/ui/src/features/ai-assistant/hooks/use-panel-resize.ts
@@ -1,0 +1,34 @@
+import { useCallback, useRef, useState } from "react";
+import { PANEL_DEFAULT_WIDTH, PANEL_MAX_WIDTH, PANEL_MIN_WIDTH } from "../constants";
+
+export function usePanelResize() {
+  const [width, setWidth] = useState(PANEL_DEFAULT_WIDTH);
+  const dragging = useRef(false);
+  const startX = useRef(0);
+  const startWidth = useRef(0);
+
+  const onMouseDown = useCallback(
+    (e: React.MouseEvent) => {
+      dragging.current = true;
+      startX.current = e.clientX;
+      startWidth.current = width;
+      e.preventDefault();
+
+      const onMouseMove = (e: MouseEvent) => {
+        if (!dragging.current) return;
+        const delta = startX.current - e.clientX;
+        setWidth(Math.min(PANEL_MAX_WIDTH, Math.max(PANEL_MIN_WIDTH, startWidth.current + delta)));
+      };
+      const onMouseUp = () => {
+        dragging.current = false;
+        window.removeEventListener("mousemove", onMouseMove);
+        window.removeEventListener("mouseup", onMouseUp);
+      };
+      window.addEventListener("mousemove", onMouseMove);
+      window.addEventListener("mouseup", onMouseUp);
+    },
+    [width],
+  );
+
+  return { width, onMouseDown };
+}
